dom/media/test/test_play_promise_11.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: promise-based play() method</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">

<script>
// Name: newSrcRejectPendingPromises
// Case: change src of an element with pending promises.
// Expected result: reject all the pending promises.

let manager = new MediaTestManager;

function initTest(test, token) {
  manager.started(token);

  let element = document.createElement(getMajorMimeType(test.type));
  element.play().then(
    (result) => {
      ok(false, `${token} is resolved with ${result}.`);
    },
    (error) => {
      if (error.name == "AbortError") {
        ok(true, `${token} is rejected with ${error.name}.`);
      } else {
        ok(false, `${token} is rejected with ${error.name}.`);
      }
    }
  ).then( () => { manager.finished(token); } );
  element.src = test.name;
}

manager.runTests(gSmallTests, initTest);

</script>
